feat(driver, cust): finish place order, reject order

+++ lib/features/customers/confirm_order/models/order_model.dart
@@ -9,7 +9,7 @@ class OrderModel {
   DateTime orderDatetime;
   String note;
   List<OrderItem> orderItems;
-  double totalPrice;
+  int totalPrice;
 
   OrderModel({
     this.id = '',
