+++ lib/features/customers/restaurant_list/controllers/cart_controller.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:food_delivery_h2d/features/customers/confirm_order/controllers/order_controller.dart';
 import 'package:get/get.dart';
 import 'package:food_delivery_h2d/features/restaurants/menu_management/models/item_model.dart';
@@ -7,20 +9,20 @@ class CartController extends GetxController {
 
   var cartItems = <Item>[].obs;
   var itemQuantities = <String, int>{}.obs;
-  double deliveryFee = 15000;
+  int deliveryFee = 15000;
 
   int get totalItems {
     return itemQuantities.values.fold(0, (sum, quantity) => sum + quantity);
   }
 
-  double get totalPrice {
+  int get totalPrice {
     return cartItems.fold(0, (sum, item) {
       var quantity = itemQuantities[item.itemName] ?? 1;
       return sum + (item.price * quantity);
     });
   }
 
-  double get orderPrice {
+  int get orderPrice {
     OrderController.instance.order.totalPrice = totalPrice;
     return totalPrice + deliveryFee;
   }
